+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+##2014-11-15 - Release 1.1.2
+###Summary
+- rename parameter manage_service to service_manage
+- Fix Issue #30 Add params for stash::service
 ##2014-10-26 - Release 1.1.1
 ###Summary
 - Replace mkrakowitzer-deploy with nanliu-staging as the default to deploy the install file.